Refactor : comment 인터페이스와 dto의 parentId 타입 Long으로 변경
#179

+++ src/main/java/com/example/mssaem_backend/global/common/Comment.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 public interface Comment {
     Long getId();
-    Integer getParentId();
+    Long getParentId();
     String getContent();
     Long getLikeCount();
     LocalDateTime getCreatedAt();
